Karma to Jest: exporter
